+++ un_yaml/un_yaml.py
@@ -12,7 +12,7 @@ class UnYaml:
     REF_ERROR = f"Value for Key {REF} does not start with {PREFIX}"
 
     @staticmethod
-    def load_yaml(filename: str, pkg: str, sub: str = "") -> dict:
+    def LoadYaml(filename: str, pkg: str, sub: str = "") -> dict:
         yaml_dir = resources.files(pkg)
         if len(sub) > 0:
             yaml_dir = yaml_dir / sub
@@ -22,8 +22,9 @@ class UnYaml:
         return yaml_data
 
     def __init__(self, yaml_data: dict) -> None:
-        self.cfg = yaml_data
-        self._info = self.cfg[UnYaml.KEY]
+        self.data = yaml_data
+        assert self.data, "UnYaml: no data"
+        self._info = self.data[UnYaml.KEY]
 
     def info(self, key: str) -> Any:
         return self._info.get(key)
@@ -54,9 +55,9 @@ class UnYaml:
             return result[int(key)]
         return result.get(key)
 
-    def get(self, keys: str) -> Any:
-        result = self.cfg
-        for key in keys.split(UnYaml.SEP):
+    def get(self, keylist: str) -> Any:
+        result = self.data
+        for key in keylist.split(UnYaml.SEP):
             item = self._get(result, key)
             result = self.expand(item)