+++ un_yaml/wrapper.py
@@ -1,2 +1,8 @@
 class Wrapper:
-    pass
+    """Wrap resource methods"""
+
+    def __init__(self, data=None):
+        self.my_data = data
+
+    async def execute(self, cmd: str, argv: dict):
+        return [cmd, argv]