+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG.md
 
+## 0.2.0 (2023-06-01)
+
+- UnConf for writeable yaml files: New, put, save
+- add typecheck via mypy
+- parse URI path components as K_UPTH
+
 ## 0.1.0 (2023-05-31)
 
 - Migrate from UDC